feat: se agregó el componente para agregar datos adicionales al piechart
